fixes #354 and #315.