@marten-seemann 2017-01-09
This PR implements an ACK algorithm similar to the one Chromium is using (in TCP_ACKING mode).

Note that the transfer speed in the benchmark tests significantly decreases, on my Macbook from about 200 MB/s to 135 MB/s and on Travis from 11.5 MB/s to 10.7 MB/s. 

The reason for this is *not* a timer issue in the `Session`, but seems to be caused by us sending more ACKs for retransmittable packets: If the packets are arriving fast enough (i.e. less than 5ms delay between two packets), we're now sending an ACK for every second packet. Before, we only sent one ACK every 5ms.
Increasing `protocol.RetransmittablePacketsBeforeAck`, which is now equal to 2, and determines every how many retransmittable packets we send an ACK, to about 20 increases the benchmark to 205 MB/s.